Rename `DATABUILDER_BRANCH` to `EHRQL_BRANCH`
