+++ INSTALL.md
@@ -17,10 +17,10 @@ This should install everything required.
 The Data Builder documentation is imported from the
 [databuilder](https://github.com/opensafely-core/databuilder) repo using the
 [mkdocs-multirepo-plugin](https://github.com/jdoiro3/mkdocs-multirepo-plugin) and built alongside the docs in this repo.  By default it uses the main branch.  This can
-be configured with the `DATABUILDER_BRANCH` environment variable:
+be configured with the `EHRQL_BRANCH` environment variable:
 
 ```
-DATABUILDER_BRANCH=my-branch just run
+EHRQL_BRANCH=my-branch just run
 ```
